DSS-1444: Use long for file size.
git-svn-id: https://svn.cesecore.eu/svn/signserver-restricted/trunk@8680 91c8134f-2a44-0410-ba45-dce2f532dbc6

+++ signserver/modules/SignServer-Client-CLI/src/main/java/org/signserver/client/cli/defaultimpl/InputSource.java
@@ -23,17 +23,17 @@ import java.io.InputStream;
  */
 public class InputSource {
     private final InputStream inputStream;
-    private final int size;
+    private final long size;
     private final String fileName;
     
-    public InputSource(final InputStream inputStream, final int size,
+    public InputSource(final InputStream inputStream, final long size,
                        final String fileName) {
         this.inputStream = inputStream;
         this.size = size;
         this.fileName = fileName;
     }
     
-    public InputSource(final InputStream inputStream, final int size) {
+    public InputSource(final InputStream inputStream, final long size) {
         this(inputStream, size, null);
     }
     
@@ -41,7 +41,7 @@ public class InputSource {
         return inputStream;
     }
     
-    public int getSize() {
+    public long getSize() {
         return size;
     }
